a final once-over
